fix: satisfy pkg/config file-count limit for CI
Colocate split-layout replace in layout.go and move split-replace tests to
pkg/config/splitreplacetests so pkg/config stays within the 15-file gate.
