+++ pkg/config/layout.go
@@ -731,3 +731,160 @@ func workstationDefForExpansion(workstationCfg interfaces.FactoryWorkstationConf
 	return def, promptFileContent
 }
 
+type factorySplitLayoutReplaceHooks struct {
+	afterStageWrite func(stagingDir string) error
+}
+
+// FactorySplitLayoutReplaceResult holds rollback and backup-discard callbacks
+// returned by ReplaceFactorySplitLayout after a successful commit.
+type FactorySplitLayoutReplaceResult struct {
+	Restore       func()
+	DiscardBackup func()
+}
+
+// ReplaceFactorySplitLayout atomically replaces an existing factory directory at
+// targetDir with a split-layout materialization of canonicalFactoryJSON. Call
+// Restore to reinstate the pre-replace tree when downstream steps fail; call
+// DiscardBackup after successful activation to remove the on-disk backup.
+func ReplaceFactorySplitLayout(targetDir string, canonicalFactoryJSON []byte) (*FactorySplitLayoutReplaceResult, error) {
+	return replaceFactorySplitLayout(targetDir, canonicalFactoryJSON, factorySplitLayoutReplaceHooks{})
+}
+
+// ReplaceFactorySplitLayoutWithAfterStageHook runs split-layout replace and invokes
+// afterStageWrite on the staged directory before validation. It exists for tests
+// in pkg/config/splitreplacetests that assert validation failures leave disk unchanged.
+func ReplaceFactorySplitLayoutWithAfterStageHook(
+	targetDir string,
+	canonicalFactoryJSON []byte,
+	afterStageWrite func(stagingDir string) error,
+) (*FactorySplitLayoutReplaceResult, error) {
+	hooks := factorySplitLayoutReplaceHooks{}
+	if afterStageWrite != nil {
+		hooks.afterStageWrite = afterStageWrite
+	}
+	return replaceFactorySplitLayout(targetDir, canonicalFactoryJSON, hooks)
+}
+
+func replaceFactorySplitLayout(targetDir string, canonicalFactoryJSON []byte, hooks factorySplitLayoutReplaceHooks) (*FactorySplitLayoutReplaceResult, error) {
+	if strings.TrimSpace(targetDir) == "" {
+		return nil, fmt.Errorf("factory directory is required")
+	}
+	if err := requireFactoryConfig(targetDir); err != nil {
+		return nil, fmt.Errorf("replace factory split layout: %w", err)
+	}
+
+	segment := filepath.Base(targetDir)
+	parentDir := filepath.Dir(targetDir)
+	stagingDir, cleanupStaging, err := stageFactorySplitLayoutReplace(targetDir, segment, parentDir, canonicalFactoryJSON, hooks)
+	if err != nil {
+		return nil, err
+	}
+	defer cleanupStaging()
+
+	backupDir, err := commitFactorySplitLayoutReplace(parentDir, targetDir, stagingDir, segment)
+	if err != nil {
+		return nil, err
+	}
+
+	return &FactorySplitLayoutReplaceResult{
+		Restore: func() {
+			restoreFactorySplitLayoutReplace(targetDir, backupDir)
+		},
+		DiscardBackup: func() {
+			_ = os.RemoveAll(backupDir)
+		},
+	}, nil
+}
+
+func stageFactorySplitLayoutReplace(
+	targetDir, segment, parentDir string,
+	canonicalFactoryJSON []byte,
+	hooks factorySplitLayoutReplaceHooks,
+) (stagingDir string, cleanup func(), err error) {
+	factoryCfg, canonical, err := normalizeNamedFactoryPayload(segment, canonicalFactoryJSON)
+	if err != nil {
+		return "", func() {}, err
+	}
+
+	sourcePath := filepath.Join(targetDir, interfaces.FactoryConfigFile)
+	stagingDir, err = os.MkdirTemp(parentDir, "."+segment+".staging-")
+	if err != nil {
+		return "", func() {}, fmt.Errorf("create staging directory for factory %q: %w", segment, err)
+	}
+	cleanup = func() {
+		_ = os.RemoveAll(stagingDir)
+	}
+
+	if _, err := writeNamedFactoryLayout(stagingDir, factoryCfg, canonical, sourcePath); err != nil {
+		return "", cleanup, fmt.Errorf("%w: %w", ErrInvalidNamedFactory, err)
+	}
+	if hooks.afterStageWrite != nil {
+		if err := hooks.afterStageWrite(stagingDir); err != nil {
+			return "", cleanup, fmt.Errorf("prepare staged factory %q: %w", segment, err)
+		}
+	}
+	if _, err := LoadRuntimeConfig(stagingDir, nil); err != nil {
+		return "", cleanup, fmt.Errorf("%w: validate factory %q config: %w", ErrInvalidNamedFactory, segment, err)
+	}
+
+	return stagingDir, cleanup, nil
+}
+
+func commitFactorySplitLayoutReplace(parentDir, targetDir, stagingDir, segment string) (backupDir string, err error) {
+	backupDir, err = os.MkdirTemp(parentDir, "."+segment+".previous-")
+	if err != nil {
+		return "", fmt.Errorf("prepare replacement backup for factory %q: %w", segment, err)
+	}
+	if err := os.Remove(backupDir); err != nil {
+		return "", fmt.Errorf("prepare replacement backup for factory %q: %w", segment, err)
+	}
+
+	if err := os.Rename(targetDir, backupDir); err != nil {
+		return "", fmt.Errorf("backup existing factory %q: %w", segment, err)
+	}
+	committed := false
+	defer func() {
+		if !committed {
+			if restoreErr := os.Rename(backupDir, targetDir); restoreErr != nil {
+				return
+			}
+			_ = os.RemoveAll(backupDir)
+		}
+	}()
+
+	if err := os.Rename(stagingDir, targetDir); err != nil {
+		return "", fmt.Errorf("commit factory %q: %w", segment, err)
+	}
+	committed = true
+	return backupDir, nil
+}
+
+func restoreFactorySplitLayoutReplace(targetDir, backupDir string) {
+	if strings.TrimSpace(targetDir) == "" || strings.TrimSpace(backupDir) == "" {
+		return
+	}
+	if _, err := os.Stat(backupDir); err != nil {
+		return
+	}
+
+	parentDir := filepath.Dir(targetDir)
+	segment := filepath.Base(targetDir)
+	trashDir, err := os.MkdirTemp(parentDir, "."+segment+".rollback-trash-")
+	if err != nil {
+		return
+	}
+	if err := os.Remove(trashDir); err != nil {
+		return
+	}
+
+	if err := os.Rename(targetDir, trashDir); err != nil {
+		_ = os.RemoveAll(trashDir)
+		return
+	}
+	if err := os.Rename(backupDir, targetDir); err != nil {
+		_ = os.Rename(trashDir, targetDir)
+		return
+	}
+	_ = os.RemoveAll(trashDir)
+}
+
